Extracted auth test code into a test helper
Extracted the code by @ryantriggs  in auth_test.py that knows how to
retrieve tokens for a user, so that the code is re-usable by other tests

* created the auth_helper.py helper module with a class called
Creds4Token
